+++ src/CQChartsArrow.cpp
@@ -61,11 +61,15 @@ draw(CQChartsPaintDevice *device) const
 
 void
 CQChartsArrow::
-draw(CQChartsPaintDevice *device, const PenBrush &penBrush) const
+draw(CQChartsPaintDevice *device, const PenBrush &penBrush, bool skipPen) const
 {
-  device_ = device;
+  device_  = device;
+  skipPen_ = skipPen;
 
   drawContents(penBrush);
+
+  device_  = nullptr;
+  skipPen_ = false;
 }
 
 void
@@ -785,7 +789,8 @@ drawLine(const Point &point1, const Point &point2, double width, const PenBrush 
 
   pen1.setWidthF(width);
 
-  device_->setPen(pen1);
+  if (! skipPen_)
+    device_->setPen(pen1);
 
   auto p1 = pixelToWindow(point1);
   auto p2 = pixelToWindow(point2);
@@ -805,7 +810,8 @@ drawPointLabel(const Point &point, const QString &text, bool above) const
 
   CQChartsUtil::setPen(tpen, true, tc, CQChartsAlpha(), 0.0);
 
-  device_->setPen(tpen);
+  if (! skipPen_)
+    device_->setPen(tpen);
 
   Point p1(point.x - 4, point.y    );
   Point p2(point.x + 4, point.y    );